protobuf/protoc_echo_plugin/ProtoCEchoPlugin: Add support for int32

+++ infra/syntax/ProtoParser.cpp
@@ -61,7 +61,7 @@ namespace infra
         {
             input >> byte;
 
-            result += (byte & 0x7f) << shift;                                                                           //TICS !POR#006
+            result += static_cast<uint64_t>(byte & 0x7f) << shift;                                                      //TICS !POR#006
             shift += 7;
         } while (!input.Failed() && (byte & 0x80) != 0);
